Infrastructure-Stack/AgentsMainReqs/requirementsKafLang.txt
# Core dependencies for your existing LangGraph setup
langchain
langgraph
langchain-anthropic
python-dotenv
pydantic
typing-extensions

# New dependencies for Kafka integration
kafka-python==2.0.2
psycopg2-binary==2.9.7

# Optional: SendGrid integration
sendgrid==6.10.0

# Optional: Additional utilities
requests==2.31.0


